Move stag headers to include directory
