+++ include/QuadDetector.h
